Make test data and fixtures available to all modules' tests
